pyproject.toml
[project]
name = "motra"
version = "0.1.0"
description = "manages tests and data collection for the motra testbed"
authors = [
    {name = "Peter Heller <peter2.heller@oth-regensburg.de>"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
license-files = [
    "*LICENSE",
]
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "python-statemachine[diagrams] (>=2.5.0,<3.0.0)",
    "pydot (>=4.0.1,<5.0.0)",
    "graphviz (>=0.21,<0.22)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "typer (>=0.19.2,<0.20.0)",
    "sh (>=2.2.2,<3.0.0)",
]
packages = [
    {include = "motra"},
    {from = "src"},
]

[project.scripts]
motra = "motra.cli.cli:motra_cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.setuptools.packages.find]
# Automatically find all packages (directories with __init__.py)
where = ["."]